Change Character
From the Main Menu, press play, and then you come to ChoseCharacter Scene to chose the character. For now its only two but you can add characters by including the character information in the database. You need the Character name, the sprite and the animation override controller. You can create an animation override controller from the animation controllers in the project. After chosing the character, you can play as the character in the game.